:recycle: Split grabber module to parser+scrapper
