+++ tests/test_scrapper.py
@@ -2,34 +2,10 @@ from asyncio import Semaphore
 
 import pytest
 
-from tj_feed.grabber import parse_comment, Comment, fetch_comments_page, fetch_last_comments, combine_batches_back, API_COMMENTS_PER_PAGE, combine_batches_forward, _fetch_page, \
-    fetch_comments_page_next_offset, search_local_max_offset, search_actual_offset
-from tj_feed.storage import create_conn
-
-
-def test_comment_parser():
-    source = {"id": 265962, "parent_id": None, "date_added": "2020-10-15T11:22:08.210391+03:00", "is_edited": False, "level": 0,
-              "text": "рандомный текст",
-              "author": {"id": 46794, "name": "Мария Илюхина", "badge": None, "participation": [], "image": "https://sun9--w0&ava=1", "ban": None},
-              "rating": {"likes": 1, "dislikes": 17, "user_vote": 0}, "status": "visible", "article_path": "/diary-businessman-podmoskovie-920k/",
-              "article_title": "Как живет предприниматель в Подмосковье с доходом 920 000 ₽", "ban": None, "image": None}
-
-    result = parse_comment(source)
-
-    assert isinstance(result, Comment)
-    assert result.comment_date == '2020-10-15T11:22:08.210391+03:00'
-    assert result.comment_content == 'рандомный текст'
-    assert result.comment_id == 265962
-    assert result.comment_rating == 1-17
-    assert result.comment_link == 'https://journal.tinkoff.ru/diary-businessman-podmoskovie-920k/#c265962'
-
-    assert result.article_path == '/diary-businessman-podmoskovie-920k/'
-    assert result.article_title == 'Как живет предприниматель в Подмосковье с доходом 920 000 ₽'
-
-    assert result.user_id == 46794
-    assert result.user_name == 'Мария Илюхина'
-    assert result.user_link == 'https://journal.tinkoff.ru/user46794/'
-    assert result.user_image == 'https://sun9--w0&ava=1'
+from tj_feed.grabber.parser import Comment
+from tj_feed.grabber.scrapper import (fetch_comments_page, fetch_last_comments, combine_batches_back, API_COMMENTS_PER_PAGE,
+                                      combine_batches_forward, _fetch_page, fetch_comments_page_next_offset, search_local_max_offset,
+                                      search_actual_offset)
 
 
 @pytest.mark.asyncio
@@ -45,7 +21,7 @@ async def test_fetch_comments_page():
 
 @pytest.mark.asyncio
 async def test_search_local_max_offset(mocker):
-    mocker.patch('tj_feed.grabber.combine_batches_forward', lambda x: [(x, 1), (x + 100, 1)])
+    mocker.patch('tj_feed.grabber.scrapper.combine_batches_forward', lambda x: [(x, 1), (x + 100, 1)])
 
     result = await search_local_max_offset(456)
 
@@ -54,7 +30,7 @@ async def test_search_local_max_offset(mocker):
 
 @pytest.mark.asyncio
 async def test_search_actual_offset(mocker):
-    mocker.patch('tj_feed.grabber.search_local_max_offset', return_value=1457)
+    mocker.patch('tj_feed.grabber.scrapper.search_local_max_offset', return_value=1457)
 
     result = await search_actual_offset(456)
 
@@ -63,11 +39,11 @@ async def test_search_actual_offset(mocker):
 
 @pytest.mark.asyncio
 async def test_fetch_comments_page_next_offset(mocker):
-    mocker.patch('tj_feed.grabber._fetch_page', return_value=dict())
+    mocker.patch('tj_feed.grabber.scrapper._fetch_page', return_value=dict())
     result = await fetch_comments_page_next_offset(2, 7, Semaphore(1))
     assert result == 0
 
-    mocker.patch('tj_feed.grabber._fetch_page', return_value=dict(data=[1, 2, 3]))
+    mocker.patch('tj_feed.grabber.scrapper._fetch_page', return_value=dict(data=[1, 2, 3]))
     result = await fetch_comments_page_next_offset(2, 7, Semaphore(1))
     assert result == 7 + 3
