+++ tests/test_webapp.py
@@ -3,7 +3,7 @@ import pytest
 
 @pytest.fixture(autouse=True)
 def mocked_grabber(mocker):
-    mocker.patch('tj_feed.grabber._fetch_page', return_value=dict())
+    mocker.patch('tj_feed.grabber.scrapper._fetch_page', return_value=dict())
 
 
 def test_index_endpoint(client):
