+++ tj_feed/grabber/scrapper.py
@@ -1,77 +1,23 @@
 import asyncio
 import logging
 import operator
-import unicodedata
 from asyncio import Semaphore
-from dataclasses import dataclass
 from itertools import chain
 from typing import Iterable, Iterator, List, Optional, Tuple
 
 import aiohttp
-from typed_json_dataclass import TypedJsonMixin
 
 from tj_feed import storage
+from tj_feed.grabber.parser import Comment, parse_comment
 
-HOST = 'https://journal.tinkoff.ru'
 API_HOST = 'https://social.journal.tinkoff.ru'
 API_ENDPOINT_COMMENTS = '/api/v13/comments/'
-
-DEFAULT_AVATAR = 'https://static2.tinkoffjournal.ru/mercury-front/fbfb8fb7b8ce70bf9516d9028e231419853c5444/face-08.982a.png'
 
 CONNECTIONS_POOL = Semaphore(5)
 CONNECTIONS_TIMEOUT = 20
 CONNECTIONS_DNS_CACHE = 300
 
 API_COMMENTS_PER_PAGE = 2000
-
-
-@dataclass
-class Comment(TypedJsonMixin):
-    user_id: int
-    user_name: str
-    user_image: Optional[str]
-
-    comment_id: int
-    comment_content: str
-    comment_date: str
-    comment_rating: int
-
-    article_path: str
-    article_title: str
-
-    @property
-    def comment_link(self) -> str:
-        return f'{HOST}{self.article_path}#c{str(self.comment_id)}'
-
-    @property
-    def user_link(self) -> str:
-        return f'{HOST}/user{str(self.user_id)}/'
-
-
-def unicode_normalize(source: str) -> str:
-    try:
-        source = source.replace('\xa0', ' ')
-    except AttributeError:
-        return ''
-
-    filtered_chars = filter(lambda char: unicodedata.category(char)[0] != 'C', source)
-    return ''.join(filtered_chars)
-
-
-def parse_comment(comment: dict) -> Comment:
-    user: dict = comment.get('author', {})
-    rating: dict = comment.get('rating', {})
-    return Comment(
-        user_id=int(user.get('id')),
-        user_name=unicode_normalize(user.get('name')),
-        user_image=user.get('image') if user.get('image') else DEFAULT_AVATAR,
-        comment_id=int(comment.get('id')),
-        comment_content=unicode_normalize(comment.get('text')),
-        comment_date=comment.get('date_added'),
-        comment_rating=rating.get('likes', 0) - rating.get('dislikes', 0),
-        article_title=unicode_normalize(comment.get('article_title')),
-        article_path=comment.get('article_path'),
-    )
 
 
 def combine_batches_back(total_limit: int, max_offset: int) -> Iterator[Tuple]:
